# src/docs/swagger.json
{
	"openapi": "3.0.0",
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"ApiErrorResponse": {
				"properties": {
					"success": {
						"type": "boolean",
						"enum": [
							false
						],
						"nullable": false
					},
					"error": {
						"type": "string"
					},
					"data": {
						"type": "number",
						"enum": [
							null
						],
						"nullable": true
					}
				},
				"required": [
					"success",
					"error"
				],
				"type": "object",
				"additionalProperties": true
			},
			"UserLoginResponse": {
				"properties": {
					"token": {
						"type": "string"
					},
					"message": {
						"type": "string"
					}
				},
				"required": [
					"token",
					"message"
				],
				"type": "object",
				"additionalProperties": true
			},
			"ApiSuccessResponse_UserLoginResponse_": {
				"properties": {
					"success": {
						"type": "boolean",
						"enum": [
							true
						],
						"nullable": false
					},
					"data": {
						"$ref": "#/components/schemas/UserLoginResponse"
					}
				},
				"required": [
					"success"
				],
				"type": "object",
				"additionalProperties": true
			},
			"UserLoginRequest": {
				"properties": {
					"username": {
						"type": "string"
					},
					"password": {
						"type": "string"
					}
				},
				"required": [
					"username",
					"password"
				],
				"type": "object",
				"additionalProperties": true
			},
			"UserRegisterRequest": {
				"properties": {
					"email": {
						"type": "string"
					},
					"username": {
						"type": "string"
					},
					"password": {
						"type": "string"
					}
				},
				"required": [
					"email",
					"username",
					"password"
				],
				"type": "object",
				"additionalProperties": true
			},
			"IUser": {
				"properties": {
					"email": {
						"type": "string"
					},
					"username": {
						"type": "string"
					},
					"password": {
						"type": "string"
					}
				},
				"required": [
					"email",
					"username",
					"password"
				],
				"type": "object",
				"additionalProperties": true
			},
			"ApiSuccessResponse_IUser_": {
				"properties": {
					"success": {
						"type": "boolean",
						"enum": [
							true
						],
						"nullable": false
					},
					"data": {
						"$ref": "#/components/schemas/IUser"
					}
				},
				"required": [
					"success"
				],
				"type": "object",
				"additionalProperties": true
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "zoro-api",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"paths": {
		"/user/login": {
			"post": {
				"operationId": "Login",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/ApiErrorResponse"
										},
										{
											"$ref": "#/components/schemas/ApiSuccessResponse_UserLoginResponse_"
										}
									]
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized"
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UserLoginRequest"
							}
						}
					}
				}
			}
		},
		"/user/register": {
			"post": {
				"operationId": "Register",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/ApiErrorResponse"
										},
										{
											"additionalProperties": true,
											"type": "object"
										}
									]
								}
							}
						}
					},
					"201": {
						"description": "Success"
					},
					"400": {
						"description": "Bad Request"
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UserRegisterRequest"
							}
						}
					}
				}
			}
		},
		"/user/{username}": {
			"get": {
				"operationId": "UserDetails",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/ApiSuccessResponse_IUser_"
										},
										{
											"$ref": "#/components/schemas/ApiErrorResponse"
										}
									]
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized"
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "username",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}